==> 0_a_ColRowFilter.ipynb <==
!pip install --upgrade ipywidgets
!pip install --upgrade pip

from utils import base_configs, deps
from utils.helpers import dir_helpers, rw_csv_helpers

import sys
import import_ipynb
import pandas as pd
import numpy as np

#### READ HINTS6 AND HINTS7 Data

base = [
    "FreqGoProvider", "GeneralHealth",
    "MedConditions_Diabetes", "MedConditions_HighBP",
    "MedConditions_HeartCondition", "MedConditions_LungDisease",
    "MedConditions_Depression",
    "AverageTimeSitting", "EverHadCancer", "Age",
    "BMI", "smokeStat",
    "WeeklyMinutesModerateExercise", "AvgDrinksPerWeek",
]

hints6_only = ["Deaf", "AverageSleepNight", "BirthGender", "PHQ4", "eCigUse"]
hints7_only = ["HealthLimits_Deaf", "BirthSex", "phq4", "ECigUse"]

# hints6_required = base + hints6_only
# hints7_required = base + hints7_only

# Alternatively,
hints6_required = base.copy()
hints6_required.extend(hints6_only)

hints7_required = base.copy()
hints7_required.extend(hints7_only)

import pandas as pd

# Read data
# df_hints6 = pd.read_excel("ip/hints6_public.xlsx", sheet_name="hints6_public", usecols=hints6_required)
# df_hints7 = pd.read_excel("ip/hints7_public.xlsx", sheet_name="hints7_public", usecols=hints7_required)
df_hints6 = rw_csv_helpers.read_csv_file("ip/0_ip_data/hints6_public.csv", columns = hints6_required, verbose = 1)
df_hints7 = rw_csv_helpers.read_csv_file("ip/0_ip_data/hints7_public.csv", columns = hints7_required, verbose = 1)

# 1) Remove AverageSleepNight from HINTS 6
df_hints6 = df_hints6.drop(columns=["AverageSleepNight"])

# 2) Rename columns in HINTS 7
df_hints7 = df_hints7.rename(
    columns={
        "HealthLimits_Deaf": "Deaf",
        "BirthSex": "BirthGender",
        "phq4": "PHQ4",
        "ECigUse": "eCigUse",
    }
)

# 3) Union (row-wise concatenate) HINTS 6 and HINTS 7
df_hints6_7_union = pd.concat([df_hints6, df_hints7], ignore_index=True)

# 4) Write to CSV
# df_union.to_csv("ip/combined_6_7/hints6_7_union.csv", index=False)

csv_path = rw_csv_helpers.write_csv_file("ip/colFiltered/hints6_colFiltered.csv", df_hints6, verbose = 1)
csv_path = rw_csv_helpers.write_csv_file("ip/colFiltered/hints7_colFiltered.csv", df_hints7, verbose = 1)

csv_path = rw_csv_helpers.write_csv_file("ip/colFiltered/hints6_7_union_colFiltered.csv", df_hints6_7_union, verbose = 1)
