==> 0_b_dataproc_imputed.ipynb <==
from utils import base_configs, deps, tr_va_te_split
from utils.helpers import dir_helpers, rw_csv_helpers, feature_distr_helpers

from utils import catboost_lib

import sys
import import_ipynb
import pandas as pd
import numpy as np

### 0 Data load

import pandas as pd

df_orig = rw_csv_helpers.read_csv_file("ip/hints6_public_filtered_v1.csv", verbose = 1)

### 1 Impute data

df_imputed = catboost_lib.impute_negatives_gpu_catboost_fast(df_orig)
df_imputed.shape

### 3 Save data

csv_path = _config.write_csv_file("op/hints6_public_filtered_v1_imputed.csv", df_imputed, verbose = 1)
